Convert back to OTP app the umbrella project
