--- apps/exalice/config/config.exs
@@ -1,8 +1,0 @@
-use Mix.Config
-
-import_config "#{Mix.env}.exs"
-
-
-config :exalice,
-  provider: ExAlice.Geocoder.Providers.Elastic,
-  geocoder: ExAlice.Geocoder.Providers.GoogleMaps
