+++ config/config.exs
@@ -1,3 +1,8 @@
 use Mix.Config
 
-import_config "../apps/*/config/config.exs"
+import_config "#{Mix.env}.exs"
+
+
+config :exalice,
+  provider: ExAlice.Geocoder.Providers.Elastic,
+  geocoder: ExAlice.Geocoder.Providers.GoogleMaps
