+++ lib/exalice/providers/elastic/elastic.ex
