+++ lib/exalice/stream_runner.ex
